+++ pkg/porter/run_test.go
@@ -15,7 +15,6 @@ import (
 	"github.com/deislabs/porter/pkg/config"
 	"github.com/deislabs/porter/pkg/context"
 	output "github.com/deislabs/porter/pkg/outputs"
-
 	"github.com/deislabs/cnab-go/bundle/definition"
 )
 
@@ -108,8 +107,10 @@ func TestActionInput_MarshalYAML(t *testing.T) {
 
 func TestApplyBundleOutputs_None(t *testing.T) {
 	p := NewTestPorter(t)
-	p.Manifest = &config.Manifest{
-		Name: "mybun",
+	p.Manifest = &config.RuntimeManifest{
+		Manifest: &config.Manifest{
+			Name: "mybun",
+		},
 	}
 	opts := NewRunOptions(p.Config)
 
@@ -124,22 +125,24 @@ func TestApplyBundleOutputs_None(t *testing.T) {
 
 func TestApplyBundleOutputs_Some_Match(t *testing.T) {
 	p := NewTestPorter(t)
-	p.Manifest = &config.Manifest{
-		Name: "mybun",
-		Outputs: []config.OutputDefinition{
-			{
-				Name: "foo",
+	p.Manifest = &config.RuntimeManifest{
+		Manifest: &config.Manifest{
+			Name: "mybun",
+			Outputs: []config.OutputDefinition{
+				{
+					Name: "foo",
 				Schema: definition.Schema{
-					Type: "string",
+						Type: "string",
+					},
+					Sensitive: true,
 				},
-				Sensitive: true,
-			},
-			{
-				Name: "123",
+				{
+					Name: "123",
 				Schema: definition.Schema{
-					Type: "string",
+						Type: "string",
+					},
+					Sensitive: false,
 				},
-				Sensitive: false,
 			},
 		},
 	}
@@ -182,14 +185,16 @@ func TestApplyBundleOutputs_Some_Match(t *testing.T) {
 
 func TestApplyBundleOutputs_Some_NoMatch(t *testing.T) {
 	p := NewTestPorter(t)
-	p.Manifest = &config.Manifest{
-		Name: "mybun",
-		Outputs: []config.OutputDefinition{
-			{
-				Name: "bar",
-			},
-			{
-				Name: "456",
+	p.Manifest = &config.RuntimeManifest{
+		Manifest: &config.Manifest{
+			Name: "mybun",
+			Outputs: []config.OutputDefinition{
+				{
+					Name: "bar",
+				},
+				{
+					Name: "456",
+				},
 			},
 		},
 	}
@@ -213,24 +218,26 @@ func TestApplyBundleOutputs_Some_NoMatch(t *testing.T) {
 
 func TestApplyBundleOutputs_ApplyTo_True(t *testing.T) {
 	p := NewTestPorter(t)
-	p.Manifest = &config.Manifest{
-		Name: "mybun",
-		Outputs: []config.OutputDefinition{
-			{
-				Name: "foo",
-				ApplyTo: []string{
-					"upgrade",
+	p.Manifest = &config.RuntimeManifest{
+		Manifest: &config.Manifest{
+			Name: "mybun",
+			Outputs: []config.OutputDefinition{
+				{
+					Name: "foo",
+					ApplyTo: []string{
+						"upgrade",
+					},
 				},
-			},
-			{
-				Name: "123",
-				ApplyTo: []string{
-					"install",
-				},
+				{
+					Name: "123",
+					ApplyTo: []string{
+						"install",
+					},
 				Schema: definition.Schema{
-					Type: "string",
+						Type: "string",
+					},
+					Sensitive: false,
 				},
-				Sensitive: false,
 			},
 		},
 	}
